refactor: extract transitions module (#793)
Move all transition methods from DPPlanner to custom_components/localshift/engine/transitions.py:
- transition(), _transition_hold(), _transition_hold_surplus(), _transition_hold_deficit()
- _transition_charge_grid(), _charge_grid_with_solar(), _charge_grid_with_deficit()
- _clip_charge_to_max_soc(), _transition_export()

Update core.py call sites and test imports to use standalone functions.
